Split Auth parser to AuthCommandParser.
